+++ lib/theme/base/app_theme/ds_app_theme.dart
@@ -76,6 +76,7 @@ class DSAppTheme {
       DSButtonThemeExtension(textTheme: _textTheme),
       DSIconButtonThemeExtension(),
       DSInputThemeExtension(),
+      DSBasicScreenFormThemeExtension(_textTheme),
     ],
   );