+++ lib/theme/components/ds_input/ds_input_theme.dart
@@ -1,5 +1,3 @@
 part of '../../ds_theme.dart';
 
-class DSInputTheme {
-  // TODO: Define theme properties for DSInput component
-}
+class DSInputTheme {}