+++ lib/theme/components/ds_input/ds_input_theme.ext.dart
@@ -5,7 +5,6 @@ enum DSInputVariants {
   secondary,
   outline,
   ghost,
-  // TODO: Define variants for DSInput component
 }
 
 class DSInputThemeExtension extends ThemeExtension<DSInputThemeExtension> {