+++ lib/theme/ds_theme.dart
@@ -16,6 +16,8 @@ part 'extensions/typography/ds_text_theme.ext.dart';
 part 'extensions/color/ds_color_theme.ext.dart';
 
 // Component Theme
+part 'components/ds_basic_screen_form/ds_basic_screen_form_theme.dart';
+part 'components/ds_basic_screen_form/ds_basic_screen_form_theme.ext.dart';
 part 'components/ds_icon_button/ds_icon_button_theme.dart';
 part 'components/ds_icon_button/ds_icon_button_theme.ext.dart';
 part 'components/ds_button/ds_button_theme.dart';